refactor: 🎨 re-org mod
